Check for network connection prior to checking updates
+ more refactoring after getting rid of FetchDownloader

+++ app/src/main/java/ua/gardenapple/itchupdater/database/installation/InstallationDao.kt
@@ -69,7 +69,7 @@ abstract class InstallationDao {
     abstract suspend fun getInstallationById(internalId: Int): Installation?
 
     @Query("SELECT * FROM $TABLE_NAME WHERE $DOWNLOAD_OR_INSTALL_ID = :downloadId AND $STATUS != $STATUS_INSTALLING")
-    abstract suspend fun getPendingInstallationByDownloadId(downloadId: Long): Installation?
+    abstract suspend fun getPendingInstallationByDownloadId(downloadId: Int): Installation?
 
     @Query("SELECT * FROM $TABLE_NAME WHERE $DOWNLOAD_OR_INSTALL_ID = :installId AND $STATUS = $STATUS_INSTALLING")
     abstract suspend fun getPendingInstallationByInstallId(installId: Long): Installation?
